Commented out on_modified, Tested in 340!
on_modified was getting called when we added directories with items inside it already (ie, copying a directory), so we were adding two folders of the same name which caused an error.  I commented out the code.  As for testing it in the lab computers, everything worked except for deleting because I couldn't sudo change the trash settings for Ubuntu.